Temp fix for double called playerChanged Event
the current fix is very hacky and need some rethinking later on. but the main problem is that after we switched from the old way to call the events and using CLib_Player als Ref Object now switch to a system where we save our Ref Object on a extra Namespace.